update to Fedora 33
Pull the Fedora 33 image from the official Fedora registry, to avoid issues with
Docker Hub throttling pulls. Also use the fedora-minimal image to reduce image
size and attack surface.

Get the Python version from an environment variable, rather than hard-coding it
in the static file path. This should make it more visible, and easier to update
when the Python version changes in the future.